Add support for multiple files in RUPS. RES-662
